remove to common and no default return value
